requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distro==1.9.0
fastapi==0.115.11
frozenlist==1.5.0
griffe==1.6.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
loguru==0.7.3
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.2.0
openai==1.68.2
openai-agents==0.0.6
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
sniffio==1.3.1
starlette==0.46.1
tavily-python==0.5.1
tiktoken==0.9.0
tomlkit==0.13.2
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
